core code updates - minor fixes
git-svn-id: http://svn.code.sf.net/p/hsqldb/svn/base@4622 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ trunk/src/org/hsqldb/QuerySpecification.java
@@ -1354,9 +1354,9 @@ public class QuerySpecification extends QueryExpression {
             Object[]        data  = new Object[indexLimitData];
             Table           table = rangeVariables[0].getTable();
             PersistentStore store = table.getRowStore(session);
-            int             count = store.elementCount(session);
+            long            count = store.elementCount(session);
 
-            data[0] = data[indexStartAggregates] = ValuePool.getInt(count);
+            data[0] = data[indexStartAggregates] = ValuePool.getLong(count);
 
             navigator.add(data);
